requirements.txt
et_xmlfile==2.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
pywin32